All pages added and most of the functionality available.

+++ app/src/main/java/com/example/jamesfarnsworthc196/ViewSelectedTerm.java
@@ -21,13 +21,15 @@ import android.widget.Toast;
 import java.util.ArrayList;
 import java.util.List;
 
+import Database.Course;
 import Database.DatabaseHelper;
+import Database.RecyclerTouchListener;
 import Database.Term;
 
 public class ViewSelectedTerm extends AppCompatActivity {
 
-    private TermsAdapter mAdapter;
     private List<Term> termsList = new ArrayList<>();
+    private List<Course> coursesList = new ArrayList<>();
     private RecyclerView recyclerView;
     private DatabaseHelper db;
     private Context context = this;
@@ -35,17 +37,44 @@ public class ViewSelectedTerm extends AppCompatActivity {
     private String endDate;
     private String termName;
     private Term term;
+    private CoursesAdapter mAdapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_view_selected_term);
+
+        recyclerView = findViewById(R.id.recyclerview);
 
         db = new DatabaseHelper(this);
 
         Intent intent = getIntent();
         int id = intent.getIntExtra("selectedTermId", 1);
         term = db.getTerm(id);
+
+
+        coursesList.addAll(db.getAllAssociatedCourses(term.getId()));
+
+        mAdapter = new CoursesAdapter(this, coursesList);
+        RecyclerView.LayoutManager mLayoutManager= new LinearLayoutManager(getApplicationContext());
+        recyclerView.setLayoutManager(mLayoutManager);
+        recyclerView.setAdapter(mAdapter);
+
+        recyclerView.addOnItemTouchListener(new RecyclerTouchListener(this, recyclerView, new RecyclerTouchListener.ClickListener(){
+            @Override
+            public void onClick(View view, final int position){
+                // load new activity here
+                // pass the database with an intent
+                Intent viewTerm = new Intent(context, ViewSelectedTermCourse.class);
+                viewTerm.putExtra("selectedCourseId", coursesList.get(position).getId());
+                int test = coursesList.get(position).getId();
+                startActivity(viewTerm);
+            }
+            @Override
+            public void onLongClick(View view, int position){
+            }
+        }));
 
 
         startDate = term.getStartdate();
@@ -112,13 +141,23 @@ public class ViewSelectedTerm extends AppCompatActivity {
                 Toast.makeText(context, "save", Toast.LENGTH_SHORT).show();
                 return true;
             case R.id.delete:
-                // delete term
-                db.deleteTerm(term);
-                // redirect to the parent activity
-                Toast.makeText(context, "delete", Toast.LENGTH_SHORT).show();
-                Intent courseIntent = new Intent(context, Terms.class);
-                startActivity(courseIntent);
-                return true;
+
+                List<Course> termCourses = new ArrayList<>();
+
+                termCourses.addAll(db.getAllAssociatedCourses(term.getId()));
+
+                if(termCourses.isEmpty()){
+                    // delete term
+                    db.deleteTerm(term);
+                    // redirect to the parent activity
+                    Toast.makeText(context, "delete", Toast.LENGTH_SHORT).show();
+                    Intent courseIntent = new Intent(context, Terms.class);
+                    startActivity(courseIntent);
+                    return true;
+                }else{
+                    Toast.makeText(context, "Courses are still assigned to this term", Toast.LENGTH_LONG).show();
+                }
+
             default:
                 return super.onOptionsItemSelected(item);
         }
